Provides `test` script for composer
Running `composer test` should execute the automatic test suite of the
project. I became fond of this workflow, of every project I have
handling that same command with the same expected outcome - and not
having to maintain a global `phpunit` command.

Travis now uses that script (`composer test`) instead of using the
global `phpunit` command that is the default behavior of that CI.

Renaming `phpunit.xml` to `phpunit.xml.dist` still makes `phpunit` to
use it as a default behavior opening the door for users to use their own
`phpunit.xml` file if they want.