Defines operands as everything after "-" or "--"
Previous definition counted everything after the first non-option
identified - which is very similar to what [getopt][1] already does.

Making the CLI parser more tolerant to what is defined as options makes
this easier to be used in projects where the command being passed is
not actually the whole responsibility of the program being used.

This **breaks compatibility** with previous versions.

[1]: https://php.net/getopt